pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "github-mcp"
version = "0.1.0"
description = "GitHub MCP Server for Claude"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
]

